refactor(blorktools): preview loading splash
