+++ packages/blorktools/src/asset_debugger/loading-splash/preview-loading-splash.js
@@ -1,0 +1,66 @@
+/**
+ * Shows loading state for a preview element
+ * @param {HTMLElement} previewElement - The preview element to show loading for
+ */
+export function showPreviewLoading(previewElement) {
+    // Create loading overlay if it doesn't exist
+    if (!previewElement.querySelector('.preview-loading')) {
+        const loadingElement = document.createElement('div');
+        loadingElement.className = 'preview-loading';
+        
+        const spinnerContainer = document.createElement('div');
+        spinnerContainer.className = 'preview-loading-spinner';
+        
+        // Create atomic spinner structure
+        const atomicSpinner = document.createElement('div');
+        atomicSpinner.className = 'atomic-spinner';
+        
+        const nucleus = document.createElement('div');
+        nucleus.className = 'nucleus';
+        
+        const orbit1 = document.createElement('div');
+        orbit1.className = 'electron-orbit';
+        const electron1 = document.createElement('div');
+        electron1.className = 'electron';
+        orbit1.appendChild(electron1);
+        
+        const orbit2 = document.createElement('div');
+        orbit2.className = 'electron-orbit';
+        const electron2 = document.createElement('div');
+        electron2.className = 'electron';
+        orbit2.appendChild(electron2);
+        
+        const orbit3 = document.createElement('div');
+        orbit3.className = 'electron-orbit';
+        const electron3 = document.createElement('div');
+        electron3.className = 'electron';
+        orbit3.appendChild(electron3);
+        
+        atomicSpinner.appendChild(nucleus);
+        atomicSpinner.appendChild(orbit1);
+        atomicSpinner.appendChild(orbit2);
+        atomicSpinner.appendChild(orbit3);
+        
+        spinnerContainer.appendChild(atomicSpinner);
+        
+        const loadingText = document.createElement('div');
+        loadingText.className = 'preview-loading-text';
+        loadingText.textContent = 'Loading...';
+        
+        loadingElement.appendChild(spinnerContainer);
+        loadingElement.appendChild(loadingText);
+        
+        previewElement.appendChild(loadingElement);
+    }
+}
+
+/**
+ * Hides loading state for a preview element
+ * @param {HTMLElement} previewElement - The preview element to hide loading for
+ */
+export function hidePreviewLoading(previewElement) {
+    const loadingElement = previewElement.querySelector('.preview-loading');
+    if (loadingElement) {
+        loadingElement.remove();
+    }
+}
